Split out watch-run and after-compile into distinct modules
